requirements.txt
boto3>=1.26.0
botocore>=1.29.0
psycopg2-binary>=2.9.5
python-json-logger>=2.0.7
tenacity>=8.2.2
typing-extensions>=4.5.0 